Move gammapy/image/profile.py -> gammapy/maps/profile.py
