+++ gammapy/image/__init__.py
@@ -1,4 +1,3 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """Sky images (2-dimensional: lon, lat)."""
 from .plotting import *
-from .profile import *
