+++ gammapy/maps/profile.py
